# release-22.2: testcat: fast stats building for inverted indexes