Backport 1/1 commits from #92974.

/cc @cockroachdb/release

---

Fixes #92906
Fixes #93006

opt tests involving inverted indexes with stats are slower than necessary because the slightly heavyweight `MakeTestingEvalContext` is called every time a scan or inverted join on the index is processed.

This change reuses the existing `eval.Context`, which also avoids a mutex lock.

Example speedup:
```
./dev test pkg/sql/opt/xform --filter TestExternal/postgis-tutorial-idx --stress
```
Without fix:
```
93 runs so far, 0 failures, over 5s
229 runs so far, 0 failures, over 10s
393 runs so far, 0 failures, over 15s
556 runs so far, 0 failures, over 20s
721 runs so far, 0 failures, over 25s
889 runs so far, 0 failures, over 30s
```
With fix:
```
148 runs so far, 0 failures, over 5s
332 runs so far, 0 failures, over 10s
510 runs so far, 0 failures, over 15s
688 runs so far, 0 failures, over 20s
871 runs so far, 0 failures, over 25s
1054 runs so far, 0 failures, over 30s
```

Release note: None

Release Justification: Fixes test-only issue